services/ai-feedback-service/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
scikit-learn==1.3.2
pandas==2.1.4
numpy==1.25.2
python-multipart==0.0.6
aiofiles==23.2.1
httpx==0.25.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
opentelemetry-api==1.21.0
opentelemetry-sdk==1.21.0
opentelemetry-instrumentation-fastapi==0.42b0
opentelemetry-instrumentation-httpx==0.42b0
opentelemetry-exporter-prometheus==1.12.0rc1
opentelemetry-exporter-jaeger==1.21.0
prometheus-client==0.19.0
structlog==23.2.0
ast-tools==0.2.0
pylint==3.0.3
bandit==1.7.5
safety==2.3.5
radon==6.0.1
mypy==1.7.1
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
httpx==0.25.2